requirements.txt
opencv-python
pyzbar
python-dotenv
psycopg2